+++ core/src/main/kotlin/dev/gumil/nookbot/repository/InMemoryOrdersRepository.kt
@@ -4,17 +4,17 @@ import dev.gumil.nookbot.entities.Order
 
 internal class InMemoryOrdersRepository : OrdersRepository {
 
-    private val ordersMap = mutableMapOf<Int, MutableList<Order>>()
+    private val ordersMap = mutableMapOf<Long, MutableList<Order>>()
 
-    override fun save(id: Int, order: Order) {
+    override fun save(id: Long, order: Order) {
         ordersMap[id]?.add(order) ?: ordersMap.put(id, mutableListOf(order))
     }
 
-    override fun getOrders(id: Int): List<Order>? {
+    override fun getOrders(id: Long): List<Order>? {
         return ordersMap[id]
     }
 
-    override fun deleteOrder(id: Int, order: Order): Boolean {
+    override fun deleteOrder(id: Long, order: Order): Boolean {
         return ordersMap[id]?.remove(order) ?: false
     }
 }
