+++ core/src/main/kotlin/dev/gumil/nookbot/repository/OrdersRepository.kt
@@ -4,9 +4,9 @@ import dev.gumil.nookbot.entities.Order
 
 interface OrdersRepository {
 
-    fun save(id: Int, order: Order)
+    fun save(id: Long, order: Order)
 
-    fun getOrders(id: Int): List<Order>?
+    fun getOrders(id: Long): List<Order>?
 
-    fun deleteOrder(id: Int, order: Order): Boolean
+    fun deleteOrder(id: Long, order: Order): Boolean
 }
