+++ core/src/main/kotlin/dev/gumil/nookbot/telegram/TelegramApi.kt
@@ -3,5 +3,5 @@ package dev.gumil.nookbot.telegram
 import dev.gumil.nookbot.entities.telegram.Update
 
 interface TelegramApi {
-    suspend fun getUpdates(offset: Int, timeout: Int): List<Update>
+    suspend fun getUpdates(offset: Long, timeout: Int): List<Update>
 }
